Add files via upload
To add images to your repo, first save the images in named files in your pc. Next, go to GitHub and open the images directory. On the top right, there is an "Add File" option. Click on that then click on "upload files". Once you've done that, choose all the files with the images that you had saved earlier and upload the then commit your changes. Don't forget to git pull to your remote device before you commit and push any new works or else you'll get an error with a long confusing message. Have fun!!